+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - New logger to send events to New Relic as NR-Logs.
 - New logger to send metrics to New Relic.
+- New logger to send traces to New Relic.
 
 ## [3.1.0] - 2021-08-11